+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 * Support of `-` as a valid filename in `upload-file` command. In future `-` will be an alias for standard input.
 * Declare official support of Python 3.12
+* Cache-Control option when uploading files
 
 ### Infrastructure
 * Remove unsupported PyPy 3.7 from tests matrix and add PyPy 3.10 instead